TestingRunner/requirements.txt
uwsgi == 2.0.18
Django == 2.1.3
django-cors-headers == 3.6.0
djangorestframework == 3.12.2
HttpRunner == 1.5.15
mysqlclient == 2.0.2
beautifulsoup4 == 4.9.3
tornado == 6.1
django-celery == 3.3.1
pymysql == 0.7.2